refactor: 시스템 및 태스크 코드 수정
- 필요없는 코드 pre_code 디렉토리로 이동
