+++ input_configuration.txt
@@ -1,8 +1,9 @@
-## Memory
-dram 1000 1   0.01   0.01
-lpm 1000 0.5  0.01  0.001
+## Simulation
+simTime 1000
+verbose 2
 
-# 메모리 타입, 용량, 일률, active 시 전력소모량, idle 시 전력소모량
+# 시뮬레이션 시간
+# verbose 0:최종결과만출력, 1:태스크실행결과출력, 2:매시간마다큐출력
 
 
 
@@ -15,4 +16,14 @@ lpm 1000 0.5  0.01  0.001
 
 # 첫 번째 줄은 core의 개수
 # 두 번째 줄 부터는 각 freq의 일률, active 전력소모량, idle 전력소모량
+# 항상 일률이 높은 것 먼저 나열되어야 함.
 
+
+
+
+## Memory
+dram 1000 1   0.01   0.01
+lpm 1000 0.5  0.01  0.001
+
+# 메모리 타입, 용량, 일률, active 시 전력소모량, idle 시 전력소모량
+# 항상 dram 먼저 나열되어야 함.
