+++ input_ga_result.txt
